fixed SessionID + bumped ResoddedFramework version to 0.6.0 to match progress

+++ src/LawnApp.h
@@ -120,7 +120,7 @@ class LawnApp : public SexyApp
 	bool mDebuggerEnabled;
 	DebuggerWindow *mDebugWindow;
 #endif
-	uint32_t mSessionID;						   //+0x87C
+	int64_t mSessionID;						   //+0x87C
 	int mPlayTimeActiveSession;			   //+0x880
 	int mPlayTimeInactiveSession;		   //+0x884
 	BoardResult mBoardResult;			   //+0x888
